requirements.txt
numpy==1.26.4
pandas==2.2.2
torch==2.2.2
yfinance
scikit-learn
matplotlib
Pillow